Removes CborPacked.
In serde_cbor 0.9, the helper function to_vec_packed created the smallest
CBOR representation of data.  When serde_cbor 0.10 was introduced, that
function was left in place and the documentation for it didn't change, but
its behavior did.  Specifically, enum variants that carry data would no
longer have the variant encoded efficiently; instead it was encoded as a
String.  As such, using to_vec_packed is not the way to get a compact
representation if your data contains enums which carry data.

Since there hasn't been a yew release yet that has CBorPacked in it,
this commit removes CborPacked.