Add detail to NodeRef docs
This switches the suggestion from using it in `update`
to using it in `rendered` within the example of a canvas
element.

I believe this is where most people would want to use it,
especially for canvas. I did not know about the `rendered`
method until I asked on Discord but I did see this page
when searching the docs site for "canvas".