Fix issues in `WebSocketTask` during connection (#1783)
During the connection handshake:

* `WebSocketTask::is_active` erronously returns `false`.
* Dropping `WebSocketTask` drops the `WebSocket` without
  calling `WebSocket::close` on it, resulting in the WebSocket
  "leaking"; it will typically continue the connection
  handshake and remain connected to the server, despite the
  client having dropped the object representing it.

Fix these by considering the task to be active if the
WebSocket `ready_state` is *either* `WebSocket::OPEN` or
`WebSocket::CONNECTING`.

Add two tests to check that a `WebSocketTask` returns `true`
for `is_active` while it's connecting, and that dropping
the `WebSocketTask` while it's connecting results in
the underlying WebSocket being closed properly.