Add event listener syntax to support custom event
Uses a similar syntax to Svelte `on:click`. Yew will continue to check
standard global event handlers and provide the respective `web_sys`
event.

For custom events the handler is registered for the event after `on:`
syntax, therefore `on:custard` will register a handler for `custard`
events. Custom event handlers will receive the `CustomEvent` type which
is a new type for `web_sys::Event` which is used to indicate that the
user is specifying a custom event handler. This should help prevent some
typo's as users expecting a specific type will get an error that Yew was
expecting the callback to be of type `Callback<CustomEvent>` where they
might have expected `Callback<MouseEvent>`.

The shorthand property assignment still works but the shorthand name
required is the name of the event and this has to be used in braces
after the listener prefix: `on:{click}`.