Refactor input and change event listeners
InputData and ChangeData have been removed as they were overly
restrictive, the implementation would only allow adding listeners to
input, textarea, or select elements and would panic at runtime
otherwise.

TypedTarget trait should help with migration for getting the value from
inputs etc. The user is much more likely to know what the target type is
and this saves Yew trying to brute force what element the target type might
be.

This is convenient enough trait but follows the same theme of JsCast
which might also be a nice way for new users to be introduced to what
JsCast can do at a limited scope.