Update examples to use TypedTarget trait
Updated to use the TypedTarget trait on event's where a value is
required from the target of an event.

Fixed instances where the `value` attribute is applied when this is not
required to do so, this reduces rendering on some examples.

Fixed using string values when floats are required, as the new approach allows for calling
HtmlInputElement::value_as_number.