Remove function components from versioned docs
function components are not part of `0.18.0` and even with the warning on
the introduction page many users still ask how they can access function
components or hooks.

The information and code blocks have fallen out of sync with the changes
on `master` so this is not even useful for users. One of the examples on
"Custom Hooks" page causes an exception in the rust markup so this issue
is resolved too and doesn't apply to the `next` version as the code
example is different.

If a user wants to learn about function components or hooks then they
can review the `next`/`master` version on yew.rs.