Fix: Hydratation of empty lists next to components. (#3630)
* fix: hydration of empty VLists

an additional (empty) VText was inserted in the second reconciliation
pass, which tried to insert itself in an invalid position. This internal
error was masked by a "fix" of the internal slot of components, which
should have still been trapped to signal that the second fixup render
was not yet run.

* fixup: new lints from clippy