Less error prone yew-router instructions (#3552)
Simply add yew-router via cargo and hope that cargo figures out the
correct version. For most users referencing git directly is wrong
and might lead to annoying errors when compiling.

Before, there was even a comment to keep the version reference up-to-date which obviously did not work and seems unpractical.