key: avoid allocation for non Rc<str> From implementations
`From<&str> for Rc` clones the content of the &str, so there are two
allocations for most types: `to_string()` and `Rc::from`.

Using a Rc<String> instead of Rc<str> internally can avoid this
additional allocation for most types, and completely avoid it for
`String`. This will now allocate when using it for Rc<str> directly
(which wasn't the case previously).

Since direct use of Rc<str> will not happen very often, I'd argue that
this tradeoff is better than the status quo.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>