Add new project to awesome.md list
This repository provides an Hello World blog use case for a full-stack Rust web application in which users can publish and read blog posts. The focus is not on the application logic but on building all the required components that a web application needs from the frontend to the database interactions.
It uses `yew`, `axum`, `mongodb` and docker compose.